fix(lsp): dummy declare variables for eclipse
Because sometimes these variables are declared later in the script
through `.vimrc.local` or other means such as `.exrc`, the `if`
expression will always be false. Thus, no language server for Java will
be registered. So, to keep preventing a complaint about undefined
variables, this simply sets them to empty strings.